feat(main-page): Add webpack and components
