+++ src/components/load.js
@@ -1,0 +1,7 @@
+/**
+ * Gets load template
+ * @return {string}
+ */
+const getLoadTemplate = () => `<button class="load-more" type="button">load more</button>`;
+
+export {getLoadTemplate};
