+++ src/components/main-container.js
@@ -1,0 +1,3 @@
+const mainContainer = document.querySelector(`.main`);
+
+export {mainContainer};
