+++ src/components/rendering.js
@@ -1,0 +1,10 @@
+/**
+ * Renders component inside container
+ * @param {Element} container
+ * @param {string} component
+ * @param {string} position for component
+ * @return {Element}
+ */
+const renderComponent = (container, component, position = `beforeend`) => container.insertAdjacentHTML(position, component);
+
+export {renderComponent};
