+++ webpack.config.js
@@ -1,0 +1,17 @@
+const path = require(`path`);
+
+module.exports = {
+  mode: `development`,
+  entry: `./src/main.js`,
+  output: {
+    path: path.join(__dirname, `public`),
+    filename: `bundle.js`
+  },
+  devtool: `source-map`,
+  devServer: {
+    contentBase: path.join(__dirname, `public`),
+    publicPath: `https://localhost:8080/`,
+    compress: true,
+    watchContentBase: true
+  }
+};
